requirements.txt
langchain_community
langchain
python-dotenv
langchain_core
langchain_groq
gradio
langchain_google_genai
groq
PyPDF2
langchain_pinecone
pinecone
sentence_transformers